requirements.txt
python-dotenv
openai
pyupbit
deep-translator